# api/openapi.yaml
openapi: 3.1.0
info:
  title: sshc Local API
  version: 0.1.0
paths:
...
  /api/v1/sync:
    get:
      operationId: getSyncStatus
      responses:
        "200":
          description: Whether a bucket is configured and what was last synced
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SyncStatus" }
        "401": { $ref: "#/components/responses/Problem" }
        "403": { $ref: "#/components/responses/Problem" }
  /api/v1/sync/settings:
    put:
      operationId: configureSync
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/SyncSettingsRequest" }
      responses:
        "200":
          description: Bucket configured for this run
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SyncStatus" }
        "400": { $ref: "#/components/responses/Problem" }
        "401": { $ref: "#/components/responses/Problem" }
        "403": { $ref: "#/components/responses/Problem" }
  /api/v1/sync/push:
    post:
      operationId: pushSnapshot
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/PassphraseRequest" }
      responses:
        "200":
          description: Snapshot written
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SyncStatus" }
        "400": { $ref: "#/components/responses/Problem" }
        "401": { $ref: "#/components/responses/Problem" }
        "403": { $ref: "#/components/responses/Problem" }
        "409": { $ref: "#/components/responses/Problem" }
  /api/v1/sync/pull:
    post:
      operationId: pullSnapshot
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/PullRequest" }
      responses:
        "200":
          description: What applying the snapshot would change, or what it applied
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PullResponse" }
        "400": { $ref: "#/components/responses/Problem" }
        "401": { $ref: "#/components/responses/Problem" }
        "403": { $ref: "#/components/responses/Problem" }
        "404": { $ref: "#/components/responses/Problem" }
        "409": { $ref: "#/components/responses/Problem" }
...
components:
  schemas:
...
    Problem:
      type: object
      additionalProperties: false
      required: [code, message]
      properties:
        code: { type: string }
        message: { type: string }
        detail: { type: string }
        path: { type: string }
        line: { type: integer }
        column: { type: integer }
        diagnostics: { type: array, items: { $ref: "#/components/schemas/Diagnostic" } }
        conflict: { $ref: "#/components/schemas/ConflictReport" }
...
    SyncDirection:
      type: string
      enum: [both, push, pull]
    SyncStatus:
      type: object
      additionalProperties: false
      required: [configured, locked, endpoint, bucket, synced, direction]
      properties:
        configured: { type: boolean }
        # フォームが空のとき、その理由。設定はマスターパスワードで封じられており、
        # それが与えられるまで読めない。アクセスキーや秘密が入ることは決してない —
        # それらは一方通行で、封をされたファイルへ向かうだけである。
        locked: { type: boolean }
        # エンドポイントとバケットは、これが何を指しているかをユーザーが見られる
        # よう返す。アクセスキーと秘密が返ることは決してない。
        endpoint: { type: string }
        bucket: { type: string }
        # すべてのオブジェクトが置かれる接頭辞。空ならバケットのルート。
        path: { type: string }
        synced: { type: boolean }
        direction: { $ref: "#/components/schemas/SyncDirection" }
        lastSyncedAt: { type: string }
        origin: { type: string }
        fileCount: { type: integer }
    SyncSettingsRequest:
      type: object
      additionalProperties: false
      required: [endpoint, bucket, accessKeyId, secretAccessKey]
      properties:
        endpoint: { type: string, minLength: 1, maxLength: 2048 }
        bucket: { type: string, minLength: 1, maxLength: 255 }
        # バケット内のどこにオブジェクトを置くか。既定の空はそのルートを指す。
        # バケットはたいていすでにこのアプリケーションにちなんで名付けられており、
        # その中で同じ名前を繰り返すフォルダは、何もない階層をひとつ増やすだけだ。
        path: { type: string, maxLength: 255 }
        region: { type: string, maxLength: 64 }
        accessKeyId: { type: string, minLength: 1, maxLength: 512 }
        secretAccessKey: { type: string, minLength: 1, maxLength: 512 }
        direction: { $ref: "#/components/schemas/SyncDirection" }
    PullRequest:
      type: object
      additionalProperties: false
      required: [passphrase]
      properties:
        passphrase: { type: string, minLength: 1, maxLength: 1024 }
        # pull は既定でプレビューする。適用は、同じパスフレーズを伴う二度目の
        # 意図的な呼び出しである。これにより、このアプリケーションが他のあらゆる
        # 書き込みの前に見せるファイル単位のプレビューを、ユーザーはここでも見る。
        apply: { type: boolean }
    SyncConflict:
      type: object
      additionalProperties: false
      required: [path, changedHere, changedThere]
      properties:
        path: { type: string }
        changedHere: { type: boolean }
        changedThere: { type: boolean }
    PullResponse:
      type: object
      additionalProperties: false
      required: [applied, conflicts, written, removed]
      properties:
        applied: { type: boolean }
        conflicts:
          type: array
          items: { $ref: "#/components/schemas/SyncConflict" }
        # パスのみ。内容を運ぶ pull のレスポンスは、秘密鍵をレスポンス本文に
        # 置くことになる。
        written:
          type: array
          items: { type: string }
        removed:
          type: array
          items: { type: string }
        origin: { type: string }
...
    PassphraseRequest:
      type: object
      additionalProperties: false
      required: [passphrase]
      properties:
        passphrase: { type: string, minLength: 1, maxLength: 1024 }
...
    Diagnostic:
      type: object
      additionalProperties: false
      required: [severity, code]
      properties:
        severity: { type: string }
        code: { type: string }
        path: { type: string }
        absolute: { type: string }
        external: { type: boolean }
        line: { type: integer }
        detail: { type: string }
...
    DiffLine:
      type: object
      additionalProperties: false
      required: [op, text]
      properties:
        op: { type: string }
        text: { type: string }
        oldLine: { type: integer }
        newLine: { type: integer }
...
    ConflictReport:
      type: object
      additionalProperties: false
      required: [path, externalChange, localChange]
      properties:
        path: { type: string }
        baseDigest: { type: string }
        diskDigest: { type: string }
        externalChange: { type: array, items: { $ref: "#/components/schemas/DiffLine" } }
        localChange: { type: array, items: { $ref: "#/components/schemas/DiffLine" } }
...
  responses:
    Problem:
      description: Request rejected
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }